Remove tic toc from benchmarker. Benchmarked func returns res, time.
The function to be benchmarked has to compute the time needed. The first
returned value is the result of the computation. The second is the
timing.

Benchmarker now exports 2 files. The one is the old file with mean, min
etc. (first run timing is excluded). The other file contains all the
individual timings, and it's name is followed by a ".timings" (first run
time included).